requirements.txt
anchor-kr==0.1.3
anitopy==2.1.1
anyio==3.6.2
asgiref==3.6.0
astroid==2.15.3
beautifulsoup4==4.9.3
better-ffmpeg-progress==2.1.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
cssselect==1.2.0
decorator==4.4.2
dill==0.3.6
dj-database-url==2.0.0
Django==4.2
django-user-agents==0.4.0
djangorestframework==3.14.0
ffmpeg-python==0.2.0
future==0.18.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.2
imageio==2.27.0
imageio-ffmpeg==0.4.8
iso8601==1.1.0
isort==5.12.0
lazy-object-proxy==1.9.0
libmagic==1.0
lxml==4.9.2
m3u8==3.4.0
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
moviepy==2.0.0.dev2
multidict==6.0.4
mysqlclient==2.1.1
numpy==1.24.2
packaging==23.1
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.2.0
proglog==0.1.10
psycopg2-binary==2.9.6
pycryptodomex==3.14.1
Pygments==2.15.0
pylint==2.17.2
pypng==0.20220715.0
pypresence==4.2.1
pySmartDL==1.3.4
python-dateutil==2.8.2
python-magic==0.4.27
python-mpv==1.0.3
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
regex==2022.10.31
requests==2.26.0
rfc3986==1.5.0
rich==13.3.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.2.1
sqlparse==0.4.3
tomlkit==0.11.7
tqdm==4.65.0
typing_extensions==4.6.1
tzdata==2023.3
ua-parser==0.16.1
urllib3==1.26.6
user-agents==2.2.0
wrapt==1.15.0
yarl==1.8.2
